data/requirements.txt
accelerate==0.30.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
auto-gptq==0.6.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
datasets==2.18.0
dill==0.3.8
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2024.2.0
gekko==1.3.0
hf-xet==1.1.5
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
optimum==1.17.1
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rouge==1.0.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.67.1
transformers==4.40.2
triton==2.2.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
